requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
flasgger==0.9.7.1
flask==2.3.3
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
mistune==3.1.3
nltk==3.9.1
numpy==2.3.0
packaging==25.0
pandas==2.2.3
pip==25.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
remla2025-team19-lib-ml @ git+https://github.com/remla2025-team19/lib-ml.git@v1.0.0
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
werkzeug==3.1.3
